decouple hashkit a bit
* Try to decouple hashkit from other code so #includes are not confused
* Move conf_pool_transform to dyn_server.c so conf is more modular

+++ src/dyn_core.c
@@ -644,7 +644,7 @@ core_loop(struct context *ctx)
     core_timeout(ctx);
     execute_expired_tasks(0);
     ctx->timeout = MIN(ctx->timeout, time_to_next_task());
-    nsd = event_wait(ctx->evb, ctx->timeout);
+    nsd = event_wait(ctx->evb, (int)ctx->timeout);
     if (nsd < 0) {
         return nsd;
     }
